refactor(api): API 엔드포인트 정의 단일 출처화 (#424)
플랫 endpoints.ts와 endpoints/ 폴더의 이중 정의로 인한 drift 위험 제거.

- 플랫 src/lib/api/endpoints.ts 삭제 (파일/폴더 이름 충돌로 barrel을 가리던 문제 해소)
- API_BASE_URL을 endpoints/base.ts 단일 정의로 통일
  (운영 URL 하드코딩 fallback 제거 → Clubs/Image/Clubnotification/ClubsBookshelf를 ./base로 이전)
- member 도메인 엔드포인트(check-nickname/additional-info/members/me/login-status)를
  endpoints/member.ts로 일원화하고, auth.ts는 auth/* 만 보유
- authService가 member 도메인 호출은 MEMBER_ENDPOINTS를 사용하도록 변경
- URL 값은 모두 동일하여 동작 변화 없음 (tsc/lint 통과)
