--- src/lib/api/endpoints.ts
@@ -1,15 +1,0 @@
-export const API_BASE_URL =
-  process.env.NEXT_PUBLIC_API_URL || "https://api.checkmo.co.kr/api/v1";
-
-export const AUTH_ENDPOINTS = {
-  LOGIN: `${API_BASE_URL}/auth/login`,
-  SIGNUP: `${API_BASE_URL}/auth/signup`,
-  EMAIL_VERIFICATION: `${API_BASE_URL}/auth/email-verification`,
-  EMAIL_CONFIRM: `${API_BASE_URL}/auth/email-verification/confirm`,
-  LOGOUT: `${API_BASE_URL}/auth/logout`,
-  ADDITIONAL_INFO: `${API_BASE_URL}/members/additional-info`,
-  CHECK_NICKNAME: `${API_BASE_URL}/members/check-nickname`,
-  PROFILE: `${API_BASE_URL}/members/me`,
-  LOGIN_STATUS: `${API_BASE_URL}/members/me/login-status`,
-  TEMP_PASSWORD: `${API_BASE_URL}/auth/temp-password`,
-};
