+++ src/lib/api/endpoints/Image.ts
@@ -1,4 +1,4 @@
-import { API_BASE_URL } from "../endpoints";
+import { API_BASE_URL } from "./base";
 
 export type ImageUploadType = "PROFILE" | "CLUB" | "NOTICE";
