+++ src/lib/api/endpoints/auth.ts
@@ -6,8 +6,5 @@ export const AUTH_ENDPOINTS = {
     EMAIL_VERIFICATION: `${API_BASE_URL}/auth/email-verification`,
     EMAIL_CONFIRM: `${API_BASE_URL}/auth/email-verification/confirm`,
     LOGOUT: `${API_BASE_URL}/auth/logout`,
-    ADDITIONAL_INFO: `${API_BASE_URL}/members/additional-info`,
-    CHECK_NICKNAME: `${API_BASE_URL}/members/check-nickname`,
-    PROFILE: `${API_BASE_URL}/members/me`,
     TEMP_PASSWORD: `${API_BASE_URL}/auth/temp-password`,
 };
